+++ loinc2hpogui/src/main/java/org/monarchinitiative/loinc2hpo/model/Model.java
@@ -36,12 +36,16 @@ public class Model {
 
     private String pathToJsonFhirFile=null;
 
+    private String pathToAutoSavedFolder = null;
+
+    private String pathToLastSession = null;
 
     /** The complete HPO ontology. */
     private HpoOntology ontology=null;
     private static final TermPrefix HPPREFIX = new ImmutableTermPrefix("HP");
     /** Key: a loinc code such as 10076-3; value: the corresponding {@link QnLoinc2HPOAnnotation} object .*/
     public Map<LoincId,UniversalLoinc2HPOAnnotation> loincAnnotationMap =new LinkedHashMap<>();
+    private Map<String, Set<LoincId>> userCreatedLoincLists = new LinkedHashMap<>();
 
     private Map<LoincId, LoincEntry> loincEntryMap;
     private HashSet<LoincId> loincIds = new HashSet<>();
@@ -53,6 +57,79 @@ public class Model {
     //private boolean tempInversed= false;
     private boolean inversedBasicMode = false; //whether inverse is checked for basic mode
     private boolean inversedAdvancedMode = false; //whether inverse is checked for advanced mode
+
+    //keep a record of Loinc lists that user searched or filtered so that user can switch back and forth
+    //key: file name used for filtering; value: a list of Loinc entries
+    private Map<String, List<LoincEntry>> filteredLoincListsMap = new LinkedHashMap<>();
+    private LinkedList<List<LoincEntry>> filteredLoincLists = new LinkedList<>();
+    private List<LoincEntry> currentLoincList;
+
+    //session changes whenever one of the following functions are called:
+    //
+    private boolean sessionChanged = false;
+
+
+    public boolean isSessionChanged() {
+        return sessionChanged;
+    }
+
+    public void setSessionChanged(boolean sessionChanged) {
+        this.sessionChanged = sessionChanged;
+    }
+
+    public void addFilteredList(String filename, List<LoincEntry> list) {
+        if (filteredLoincListsMap.containsKey(filename)) { //update sequence in map and list
+            filteredLoincListsMap.remove(filename);
+            filteredLoincLists = new LinkedList<>();
+            filteredLoincLists.addAll(filteredLoincListsMap.values());
+        }
+        filteredLoincListsMap.put(filename, list);
+        filteredLoincLists.add(list);
+        currentLoincList = list;
+    }
+
+    public List<LoincEntry> previousLoincList(){
+        if (filteredLoincLists.isEmpty()) {
+            return null;
+        }
+        int current_i = filteredLoincLists.indexOf(currentLoincList);
+        if (current_i - 1 >= 0) {
+            currentLoincList = filteredLoincLists.get(current_i - 1);
+        } else {
+            currentLoincList = filteredLoincLists.getLast();
+        }
+        return currentLoincList;
+    }
+
+    public List<LoincEntry> nextLoincList() {
+        if (filteredLoincLists.isEmpty()) {
+            return null;
+        }
+        int current_i = filteredLoincLists.indexOf(currentLoincList);
+        if (current_i + 1 <= filteredLoincLists.size() - 1) {
+            currentLoincList = filteredLoincLists.get(current_i + 1);
+        } else {
+            currentLoincList = filteredLoincLists.getFirst();
+        }
+        return currentLoincList;
+    }
+
+    public List<LoincEntry> getLoincList(String listFileName) {
+        if (filteredLoincListsMap.containsKey(listFileName)) {
+            currentLoincList = filteredLoincListsMap.get(listFileName);
+            filteredLoincListsMap.remove(listFileName);//remove and add to keep order
+            filteredLoincListsMap.put(listFileName, currentLoincList);
+            filteredLoincLists.clear();
+            filteredLoincLists.addAll(filteredLoincListsMap.values());
+            return currentLoincList;
+        } else {
+            return null;
+        }
+    }
+
+    public Map<String, List<LoincEntry>> getFilteredLoincListsMap() {
+        return this.filteredLoincListsMap;
+    }
 
     public void setLoincEntryMap(Map<LoincId, LoincEntry> map) {
         this.loincEntryMap = map;
@@ -111,6 +188,22 @@ public class Model {
     public void setPathToHpOwlFile(String p) { pathToHpoOwlFile = p;
     }
     public void setBiocuratorID(String id){biocuratorID=id;}
+
+    public String getPathToAutoSavedFolder() {
+        return pathToAutoSavedFolder;
+    }
+
+    public void setPathToAutoSavedFolder(String pathToAutoSavedFolder) {
+        this.pathToAutoSavedFolder = pathToAutoSavedFolder;
+    }
+
+    public String getPathToLastSession() {
+        return pathToLastSession;
+    }
+
+    public void setPathToLastSession(String pathToLastSession) {
+        this.pathToLastSession = pathToLastSession;
+    }
 
     public String getPathToLoincCoreTableFile() {
         return pathToLoincCoreTableFile;
@@ -202,6 +295,13 @@ public class Model {
 
     public Map<LoincId,UniversalLoinc2HPOAnnotation> getLoincAnnotationMap(){ return loincAnnotationMap; }
 
+    public Map<String, Set<LoincId>> getUserCreatedLoincLists() {
+        return userCreatedLoincLists;
+    }
+
+    public void addUserCreatedLoincList(String listName, Set<LoincId> list) {
+        this.userCreatedLoincLists.put(listName, list);
+    }
 
     private void init() {
     }
@@ -254,6 +354,12 @@ public class Model {
             if (pathToHpoOwlFile!= null) {
                 bw.write(String.format("hp-owl:%s\n", pathToHpoOwlFile));
             }
+            if (pathToAutoSavedFolder != null) {
+                bw.write(String.format("autosave to:%s\n", pathToAutoSavedFolder));
+            }
+            if (pathToLastSession != null) {
+                bw.write(String.format("last session:%s\n", pathToLastSession));
+            }
             bw.close();
         } catch (IOException e) {
             e.printStackTrace();
@@ -283,6 +389,8 @@ public class Model {
                 else if (key.equals("annotationFile")) this.pathToAnnotationFile = value;
                 else if (key.equals("hp-obo")) this.pathToHpoOboFile = value;
                 else if (key.equals("hp-owl")) this.pathToHpoOwlFile = value;
+                else if (key.equals("autosave to")) this.pathToAutoSavedFolder = value;
+                else if (key.equals("last session")) this.pathToLastSession = value;
             }
             br.close();
         } catch (IOException e) {