+++ CHANGELOG.md
@@ -49,7 +49,33 @@ Multiple features are added. This will be the baseline for future tracking.
 * Refactor pom files. 
 
 * Refactor gitignore file. 
+
+## v1.0.4
+
+* Allow adding multiple labels to Github issues
+
+* Disable the function to clear annotation fields during manual query
+
+* Loinc entries change color if they have been annotated
+
+* Add tooltips to HPO listview and treeview
+
+* Allow user to switch to previously selected Loinc list
+
+* Allow user to categorize Loinc entries
+
+## v1.1.0 
+
+* New develop version
+
+Additional changes for this version
+
+* Change menu `Edit` to `Configuration`
+- [ ] update tutorial
+
+* Create new features that allow user to manipulate a session
   
+* Automatically retrieve information from auto-saved data for last session